chore: address PR feedback
